api/requirements.txt
fastapi
uvicorn[standard]
python-multipart
PyYAML